Support performing initial setup and returning
There are special cases like rolekit where we want to be able to
determine whether the initial setup of FreeIPA succeeded before
continuing with processing or providing feedback to the user. This
allows us to run the ipa-server-configure-first routine in the
container and get a success-or-failure result back, rather than
leaving the successful initialization running indefinitely.