Add support for --build-arg VERSION=<desired-version>
In rebuilding the container, we may wish to uprade a component in the OS
image, but *not* bump forward to the latest freeipa version the upstream
OS provides.

This is semi-advanced usage.  If a user is forcing an older version, they're
responsible for the potential build failures; however, we should give
them the option to at least lock the version if they wish.

This has been validated for all Dockerfile targets except for rhel7;
I do not have access to rhel7 to test.