Create machine-id at startup
D-Bus requires a proper /etc/machine-id to work. The file
/etc/machine-id is empty by default. In a read-only container D-Bus is
unable to create machine-id in /etc/ or /var/lib/dbus/ and then fails
with a cryptic error message:

   ERROR launcher_run_child @ ../src/launch/launcher.c +325: No medium found

The init-data script now creates /data/etc/machine-id when the file
does not exist and /etc/machine-id is a symlink. By default
/etc/machine-id is a symlink to /data/etc/machine-id. Container runtimes
are still free to bind-mount a custom /etc/machine-id.

Signed-off-by: Christian Heimes <cheimes@redhat.com>