Link to the workflow run that built the image.
Sadly, https://github.com/opencontainers/image-spec/blob/main/annotations.md
does not specify a separate annotation key for "this is the environment
and job where the image was built" and org.opencontainers.image.source
seems the closes fit. The link to the actual commit and thus the
repository is of course also on that workflow run page.