It does not seem necessary to create /var/log/journal volume.
We drive the journal to /run/log (and via symlink to /data/var/log).

With podman run in systemd mode, the /var/log/journal volume gets
created automatially anyway. That's why we have to have the directory
on the image even if we do not plan to actually use it -- we avoid
the podman diff reports.